% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apng.R
\name{apng}
\alias{apng}
\title{Generate APNG file from PNG files}
\usage{
apng(input_files, output_file = "output.png", delay_num = 0L, delay_den = 100L)
}
\arguments{
\item{input_files}{paths to input PNG files}

\item{output_file}{path to output APNG file}

\item{delay_num}{the numerator of the frame delay (delay is \eqn{\frac{delay_num}{delay_den}})}

\item{delay_den}{the denominator of the frame delay (delay is \eqn{\frac{delay_num}{delay_den}})}
}
\description{
Generate APNG file from PNG files
}
